backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
python-multipart
fpdf2
python-docx